+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.58.0] 2023-11-08
+### Added
+- `pickFile`: `"display": "transform"`, use `transform` command to manipilate `value`, `label` and `description` of QuickPick list items
+### Fixed
+- <code>&dollar;{result}</code> variable
+
 ## [1.57.0] 2023-10-21
 ### Added
 - <code>&dollar;{command:<em>name</em>}</code>: property `command` can contain variables